1.ManufacturingResult is modified into a super class. Two subclasss are created :simpleManufacturingResult and ComplexManufacturingResult, each of them refer to separated tables in database; 2.Introduced command pattern to save both simpleManufacturingResult and ComplexManufacturingResult;
