+++ manufacturing-result-crud/src/main/java/com/mytech/domain/ScanResultComplex.java
