+++ manufacturing-result-crud/src/main/java/com/mytech/domain/ScanResultOtherComplex.java
